Refactor services into domain and repository layers
